start refactor of generate transmittable vlp class
