+++ spec/operations/fdsh/jobs/generate_transmittable_vlp_intial_verification_payload_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 require 'shared_examples/person_cv3'
 
-RSpec.describe Fdsh::Jobs::GenerateTransmittableVlpPayload, dbclean: :after_each do
+RSpec.describe Fdsh::Jobs::Vlp::GenerateTransmittableInitialVerificationPayload, dbclean: :after_each do
   include_context "person hash for cv3"
 
   subject { described_class.new }
